# Preload the root service picture, it's called in a loop