--- a/app/presenters/tree_builder_services.rb
+++ b/app/presenters/tree_builder_services.rb
@@ -24,7 +24,12 @@ class TreeBuilderServices < TreeBuilder
 
   # Get root nodes count/array for explorer tree
   def x_get_tree_roots(count_only, _options)
-    count_only_or_objects(count_only, rbac_filtered_objects(Service.roots), "name")
+    roots = rbac_filtered_objects(Service.roots)
+
+    # Preload the root service picture since it's called by TreeNodeBuilder#build
+    MiqPreloader.preload(roots, :picture)
+
+    count_only_or_objects(count_only, roots, "name")
   end
 
   def x_get_tree_service_kids(object, count_only)