Purpose or Intent
-----------------

Services -> My Services was taking ~47 seconds in the UI with a large database with over 9000 services.
With this change, it's down to 26 seconds.

 Each root service picture is called by TreeNodeBuilder#build.

```ruby
Benchmark.ms { TreeBuilderServices.new(:svcs_tree, :svcs, {:trees=>{:svcs_tree=>{:tree=>:svcs_tree, :type=>:svcs, :klass_name=>"TreeBuilderServices", :leaf=>"Service", :add_root=>true, :open_nodes=>[], :lazy=>true, :full_ids=>true}}}, true) }
```
Note, there might be some model autoloading skewing the console/UI timings:

Before/after with a large database (9400+ services):
In the console: 55 -> 23 seconds
In the UI on /service/explorer: 47 -> 26 seconds

With help from @NickLaMuro @Fryguy