@kbrock 2016-06-23
@jrafanie This is nice

Is there a reason not to just go with the pure rails approach?

```ruby
has_one    :picture, :through => :service_template
```

After my db is migrated (global zone), I'll see how this fairs. (hope it will have completed by morning - we need to revisit binary_blob migration)