@Fryguy 2016-06-23
@kbrock For the binary blob migration you will want to take https://github.com/ManageIQ/manageiq/pull/9119