@kbrock 2016-06-23
I think the full numbers make this work even more impressive:

FREED: 42,732,026
|        ms |    bytes |  objects |queries | query (ms) |     rows |`comments`
|       ---:|      ---:|      ---:|  ---:|      ---:|      ---:| ---
|  62,138.3 | 265,176,036 | 46,137,882 |37,927 |  7,914.1 |   37,924 |before

FREED: 20,521,546
|        ms |    bytes |  objects |queries | query (ms) |     rows |`comments`
|       ---:|      ---:|      ---:|  ---:|      ---:|      ---:| ---
|  22,220.6 | 97,646,975 | 23,925,291 |9,495 |  2,319.2 |    9,496 |after

Using miq's virtual_has_one and using an actual `has_one` is basically the same memory profile.