@jrafanie 2016-06-23
@kbrock Cool, thanks for the details.  I'm not opposed using has_one through.  Is this a drop-in replacement @kbrock?

I'm not sure why it's a virtual_has_one right now other than we have a method we want to call for the relationship.  If we can eliminate the method and use a real relationship, I'm good with that change.