@kbrock 2016-06-23
@jrafanie yes it is a drop in replacement. Ladas used it and it made me concerned. Then I realized it is basically defined for the delegation use case.