@jrafanie 2016-06-23
@kbrock makes a great point RE: removing the picture method and changing the `virtual_has_one :picture, :uses=> :service_template` to `has_one :picture, :through => :service_template`.  It seems like a win-win.  Let's use rails since it appears we don't really need a virtual has_one... We'll review this in a future PR, maybe it's obvious but there could something we're missing.