@chessbyte 2016-07-05
@jrafanie @dmetzger57 @kbrock is there a BZ for this?